+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ UNRELEASED
 ----------
 
 - Support for Python 3.2, 3.3, and 3.4 is no longer tested.
+- Add new list_users function to DevpiCommandWrapper
 
 0.4.3
 -----